FIX: compiler warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@2673 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkLogHistogram.h
@@ -120,7 +120,7 @@ public:
    */
   virtual Types::DataItem BinToValue ( const size_t bin ) const 
   {
-    return this->Superclass::BinToValue( exp( static_cast<Types::DataItem>( bin ) / (this->GetNumBins()-1) * this->m_LogNumBins ) - 1 );
+    return this->Superclass::BinToValue( static_cast<size_t>( exp( static_cast<Types::DataItem>( bin ) / (this->GetNumBins()-1) * this->m_LogNumBins ) - 1 ) );
   }
 
 protected:
